requirements.txt
annotated-types==0.7.0
anyio==4.10.0
authlib==1.6.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
google-search-results==2.4.2
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
ollama==0.5.3
openai==1.101.0
packaging==25.0
playwright==1.54.0
protobuf==6.33.0
pycparser==2.23
pydantic==2.11.7
pydantic-core==2.33.2
pyee==13.0.0
python-dotenv==1.1.1
python-magic==0.4.27
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
validators==0.35.0
weaviate-client==4.17.0
